most updates from comments
