+++ src/generateBreaks/breaks.js
@@ -8,7 +8,6 @@ function calculateClassBreaks (features, classification) {
   const values = getFieldValues(features, classification.classificationField)
   // make sure there aren't more breaks than values
   if (classification.breakCount > values.length) classification.breakCount = values.length
-
   // calculate break ranges [ [a-b], [b-c], ...] from input values
   return classifyBreaks(values, features, classification)
     .map((value, index, array) => {
@@ -18,6 +17,8 @@ function calculateClassBreaks (features, classification) {
 }
 
 function classifyBreaks (values, features, classification) {
+  // TODO: determine how to handle minValue, maxValue, & label when Quantile values are heavily skewed
+  // TODO: implement last two classification methods
   if (classification.normalizationType) values = normalizeValues(values, features, classification)
   const classifier = new Classifier()
   classifier.setSeries(values)
@@ -28,22 +29,15 @@ function classifyBreaks (values, features, classification) {
     case 'esriClassifyEqualInterval': return classifier.classify('equal_interval')
     case 'esriClassifyNaturalBreaks': return classifier.classify('jenks')
     case 'esriClassifyQuantile': return classifier.classify('quantile')
-    // TODO: determine how to handle minValue, maxValue, & label when Quantile values are heavily skewed
-    // TODO: implement last two classification methods
-    case 'esriClassifyGeometricalInterval': return undefined
-    case 'esriClassifyStandardDeviation':
-      if (classification.standardDeviationInterval) {
-        // TODO: either use a different library to classify, or find how to integrate interval into calculation
-        return classifier.classify('std_deviation')
-      } else {
-        // handle when a user doesn't add a standard deviation interval
-      } break
+    case 'esriClassifyGeometricalInterval': throw new Error('Classification method not yet supported')
+    case 'esriClassifyStandardDeviation': return classifier.classify('std_deviation')
     default: throw new Error('invalid classificationMethod: ', classMethod)
   }
 }
 
 // this should really be a built-in called: Math.bump()
 function calculateMinValue (value, index, array) {
+  // TODO: determine how to fix rounding on large decimal places (e.g., [1 - 2.33334], [2.3333*6* - 3])
   let minValue = array[index - 1] || array[0]
   if (isNaN(minValue)) throw new Error('Previous break value is non-numeric')
   if (minValue !== 0 && minValue !== array[0]) {
@@ -79,3 +73,19 @@ function calculateUniqueValue (features, classification) {
 }
 
 module.exports = { calculateClassBreaks, calculateUniqueValue }
+
+/* TODO: find way to include standardDeviationInterval
+function calculateStandardDeviation (values, classification, classifier) {
+  return classifier.classify('std_deviation')
+  if (classification.standardDeviationInterval) {
+    const interval = classification.standardDeviationInterval
+    switch (interval) {
+      case '1': return classifier.classify('std_deviation')
+      case '.5': return classifier.classify('std_deviation')
+      case '.33': return classifier.classify('std_deviation')
+      case '.25': return classifier.classify('std_deviation')
+      default: throw new Error('Invalid standardDeviationInterval. Must specify one of: 1, .5, .33, .25')
+    }
+  } else throw new Error('Must include standardDeviationInterval')
+}
+*/
